+++ mm/src/audio/code_8019AF00.c
@@ -5711,14 +5711,14 @@ void Audio_MuteBgmPlayersForFanfare(void) {
  * Sets up seqId to play on seqPlayerIndex 1
  */
 void Audio_PlayFanfare(u16 seqId) {
-    u8 prevFontBuff[16];
-    u8 fontBuff[16];
+    u16 prevFontBuff[16];
+    u16 fontBuff[16];
     u16 prevSeqId = AudioSeq_GetActiveSeqId(SEQ_PLAYER_FANFARE);
     u32 outNumFonts;
     // 2S2H [Custom Audio] `prevSeqId` and `seqId` had a `& 0xFF`. This was likely OK because we always load soundFonts
     // but it was removed anyway just in case.
-    u8* prevFontId = AudioThread_GetFontsForSequence(prevSeqId, &outNumFonts, prevFontBuff);
-    u8* fontId = AudioThread_GetFontsForSequence(seqId, &outNumFonts, fontBuff);
+    u16* prevFontId = AudioThread_GetFontsForSequence(prevSeqId, &outNumFonts, prevFontBuff);
+    u16* fontId = AudioThread_GetFontsForSequence(seqId, &outNumFonts, fontBuff);
     // BENTODO
     // #region 2S2H [Audio] TODO: Fixes fanfare crash, should/can be removed after audio is done
     // if ((prevSeqId == NA_BGM_DISABLED) || (*prevFontId == *fontId)) {
